requirements.txt
requests
pandas
numpy
openpyxl
xlsxwriter
matplotlib
seaborn
plotly
python-dotenv
fuzzywuzzy
python-levenshtein
datetime
